+++ test/tests/BlockComments.hs
@@ -9,12 +9,28 @@
     = A
 --    ^ constant.other.haskell
 
+  data B =
+--     ^ storage.type.haskell
+{- comment line not gobbled up by the preprocessor -}
+-- <~~----------------------------------------------------- comment.block.haskell
+    B
+--  ^ constant.other.haskell
+
+  newtype A =
+--        ^ storage.type.haskell
+{- comment line not gobbled up by the preprocessor
+  over multiple lines -}
+-- <~~---------------------- comment.block.haskell
+    A
+--  ^ constant.other.haskell
+
   newtype B =
 --        ^ storage.type.haskell
 {- comment line not gobbled up by the preprocessor -}
 -- <~~----------------------------------------------------- comment.block.haskell
     B
 --  ^ constant.other.haskell
+
 
   data Foo where
 {- comment line not gobbled up by the preprocessor