+++ test/tests/Haddocks.hs
@@ -10,23 +10,17 @@
     doc -}
 --  ^^^ comment.block.documentation.haskell
 
-
-g :: a -- ^ doc
---          ^^^ comment.block.documentation.haskell
-  -> b --^ notDoc
---     ^^^^^^^^^^^ - comment.block.documentation.haskell
-  -> c
-
-
-
+    {- | also doc -}
+--  ^^^^^^^^^^^^^^^^ comment.block.documentation.haskell
+ 
     -- | docs
---     ^^^^^^ - comment.block.documentation.haskell
+--     ^^^^^^ comment.block.documentation.haskell
     -- $ docs
---     ^^^^^^ - comment.block.documentation.haskell
+--     ^^^^^^ comment.block.documentation.haskell
     -- ^ docs
---     ^^^^^^ - comment.block.documentation.haskell
+--     ^^^^^^ comment.block.documentation.haskell
     -- * docs
---     ^^^^^^ - comment.block.documentation.haskell
+--     ^^^^^^ comment.block.documentation.haskell
     a --| notDocs
 --  ^^^^^^^^^^^^^ - comment.block.documentation.haskell
     a --$ notDocs
@@ -59,3 +53,18 @@ g :: a -- ^ doc
 --  ^^^^^^^^^^^^ comment.block.documentation.haskell
     {- * docs -}
 --  ^^^^^^^^^^^^ comment.block.documentation.haskell
+
+        -- | docs
+--      ^^^^^^^^^ comment.block.documentation.haskell
+        -- continuing docs
+--      ^^^^^^^^^^^^^^^^^^ comment.block.documentation.haskell
+        -- still docs
+--      ^^^^^^^^^^^^^ comment.block.documentation.haskell
+    -- no longer docs
+--     ^^^^^^^^^^^^^^ - comment.block.documentation.haskell
+
+g :: a -- ^ doc
+--          ^^^ comment.line.documentation.haskell
+  -> b --^ notDoc
+--     ^^^^^^^^^^^ - comment.line.documentation.haskell comment.block.documentation.haskell 
+  -> c