+++ test/tickets/T0150.hs
@@ -1,0 +1,9 @@
+-- SYNTAX TEST "source.haskell" "Pragma after data declaration"
+
+    data D = C
+
+    {-# COMPLETE P #-}
+--  ^^^^^^^^^^^^^^^^^^ meta.preprocessor.haskell
+--  ^^^^^^^^^^^^^^^^^^ - comment.block.haskell
+    pattern P :: A
+    pattern P = C